sdk/v2/system/api_client_test.go
package system

import (
	"context"
	"net/http"
	"net/http/httptest"
	"testing"

	rmTesting "github.com/brigadecore/brigade/sdk/v2/internal/restmachinery/testing" // nolint: lll
	"github.com/stretchr/testify/require"
)

func TestNewAPIClient(t *testing.T) {
	client := NewAPIClient(
		rmTesting.TestAPIAddress,
		rmTesting.TestAPIToken,
		nil,
	)
	require.IsType(t, &apiClient{}, client)
	rmTesting.RequireBaseClient(t, client.(*apiClient).BaseClient)
}

func TestAPIClientPing(t *testing.T) {
	server := httptest.NewServer(
		http.HandlerFunc(
			func(w http.ResponseWriter, r *http.Request) {
				require.Equal(t, http.MethodGet, r.Method)
				require.Equal(
					t,
					"/ping",
					r.URL.Path,
				)
				w.WriteHeader(http.StatusOK)
			},
		),
	)
	defer server.Close()
	client := NewAPIClient(server.URL, rmTesting.TestAPIToken, nil)
	err := client.Ping(context.Background())
	require.NoError(t, err)
}
